Refactor http.
